Move common Wavefunction members to BaseWavefunction. Tests pass
