+++ psi4/src/psi4/libmints/basewavefunction.cc
@@ -27,6 +27,7 @@
  */
 
 #include "basewavefunction.h"
+#include "wavefunction.h"
 
 #include "psi4/liboptions/liboptions.h"
 #include "psi4/libpsi4util/process.h"
