+++ psi4/src/psi4/libmints/basewavefunction.h
@@ -30,15 +30,29 @@
 #define _psi_src_lib_libmints_basewavefunction_h
 
 #include "psi4/psi4-dec.h"
+#include "psi4/libmints/dimension.h"
 
 #include <array>
+#include <cstddef>
+#include <map>
 #include <memory>
+#include <string>
+#include <tuple>
+#include <vector>
 
 namespace psi {
 
 class Molecule;
 class BasisSet;
+class IntegralFactory;
+class MintsHelper;
+class Wavefunction;
+class MatrixFactory;
 class Options;
+class SOBasisSet;
+class PCM;
+class PSIO;
+class ExternalPotential;
 
 /*! \ingroup MINTS
  *  \class BaseWavefunction
@@ -46,18 +60,107 @@ class Options;
  */
 class PSI_API BaseWavefunction {
    protected:
-    /// Molecule that this wavefunction is run on
-    std::shared_ptr<Molecule> molecule_;
+    /// Name of the wavefunction
+    std::string name_;
+
+    /// Module name for CURRENT ENERGY
+    std::string module_;
 
     /// The ORBITAL basis
     std::shared_ptr<BasisSet> basisset_;
 
+    /// Primary basis set for SO integrals
+    std::shared_ptr<SOBasisSet> sobasisset_;
+
+    /// Molecule that this wavefunction is run on
+    std::shared_ptr<Molecule> molecule_;
+
     /// Options object
     Options& options_;
+
+    // PSI file access variables
+    std::shared_ptr<PSIO> psio_;
+
+    /// Integral factory
+    std::shared_ptr<IntegralFactory> integral_;
+
+    /// MintsHelper
+    std::shared_ptr<MintsHelper> mintshelper_;
+
+    /// Matrix factory for creating standard sized matrices
+    std::shared_ptr<MatrixFactory> factory_;
+
+    std::shared_ptr<Wavefunction> reference_wavefunction_;
+
+    /// How much memory you have access to.
+    long int memory_;
+
+    /// Perturb the Hamiltonian?
+    int perturb_h_;
+    /// With what...
+    enum FieldType { nothing, dipole_x, dipole_y, dipole_z, dipole, embpot, dx, sphere };
+    FieldType dipole_field_type_;
 
     /// How big of a field perturbation to apply
     std::array<double, 3> dipole_field_strength_;
 
+    /// Debug flag
+    size_t debug_;
+    /// Print flag
+    size_t print_;
+
+    /// Total frozen core orbitals
+    int nfrzc_;
+
+    /// Number of frozen core per irrep
+    Dimension frzcpi_;
+    /// Number of frozen virtuals per irrep
+    Dimension frzvpi_;
+
+    /// Number of so per irrep
+    Dimension nsopi_;
+    /// Number of mo per irrep
+    Dimension nmopi_;
+
+    /// The energy associated with this wavefunction
+    double energy_;
+
+    /// Frozen-core energy associated with this wavefunction
+    double efzc_;
+
+    /// Total number of SOs
+    int nso_;
+    /// Total number of MOs
+    int nmo_;
+    /// Number of irreps
+    int nirrep_;
+
+    /// Should nuclear electrostatic potentials be available, they will be here
+    std::shared_ptr<std::vector<double>> esp_at_nuclei_;
+
+    /// If atomic point charges are available they will be here
+    std::shared_ptr<std::vector<double>> atomic_point_charges_;
+
+    /// Should natural orbital occupations be available, they will be here
+    std::vector<std::vector<std::tuple<double, int, int>>> no_occupations_;
+
+    // The external potential for the current wave function
+    std::shared_ptr<ExternalPotential> external_pot_;
+
+    // Collection of scalar variables
+    std::map<std::string, double> variables_;
+
+    // Collection of external potentials; this member variable is provisional and might be removed in the future.
+    // This member variable is currently used for passing ExternalPotential objects to the F/I-SAPT code
+    // The above defined external_pot_ member variable contains the total external potential defined for the current
+    // wave function. For F/I-SAPT, we need a set of external potential that can be assigned to either the interacting
+    // fragments or to the environment
+    // For F/I-SAPT the keys can be A, B, or C (all optionals), where A and B signify the interacting subsystem
+    // and C signify the envirnoment
+    std::map<std::string, std::shared_ptr<ExternalPotential>> potentials_;
+
+    // Polarizable continuum model
+    std::shared_ptr<PCM> PCM_;
     /// Polarizable continuum model enabled?
     bool PCM_enabled_;
